requirements.txt
streamlit==1.20.0
pandas==1.5.2
Pillow==9.3.0
mutagen==1.46.0
ExifRead==3.0.0
PyPDF2==3.0.1
docx2txt==0.8

